Not filtering out the NULL and LGND on URL call
This was calling the METAR URL with all the NULL and LGND values still in it. 

Repeating code like this should be in functions or even classes. At least this means I can stretch my coding skills once again and dive through all this code. :)

Love the project at this point!!